batch type now being updated from AssignForce
